More DCF refactoring
